refactor(desktop): extract sweep.rs, add name pre-filter and shared helpers
Move the exact-path harness sweep into runtime/sweep.rs so runtime.rs
stays within its file-size budget (reverts the override bump). Extract
two shared low-level helpers — collect_all_pids() and procargs2_buffer() —
and convert all four duplicate proc_listallpids loop sites (sweep_system_
agent_processes, collect_same_instance_orphans, reap_dead_instance_agents,
desktop_is_alive_for_instance) and the two duplicate KERN_PROCARGS2 two-phase
sysctl sites (process_has_buzz_marker, extract_buzz_marker_value) to use them.

Add a cheap proc_name pre-filter in collect_process_snapshots before the
expensive KERN_PROCARGS2 exe-path fetch — mirrors the existing pattern in
sweep_system_agent_processes and cuts the expensive two-sysctl calls by
~99.9%. Strip " (deleted)" from Linux /proc/<pid>/exe paths so stale orphans
whose binary was replaced still match. Canonicalize expected_harness_exe_path
with a raw-path fallback so symlinked bundles compare correctly. Unify the
skip_pids calling convention to &[u32] end-to-end. Add a note at the
restore.rs call site about future consolidation to a single shared snapshot.
